split overlays into separate files
